修复潜在的 long 转 int 问题

+++ common-util/src/main/java/com/magicliang/transaction/sys/common/util/DateUtils.java
@@ -681,7 +681,7 @@ public class DateUtils {
     }
 
     public static int differentDaysByMillisecond(long startTime, long afterTime) {
-        int days = (int) ((afterTime - startTime) / (1000 * 3600 * 24)) + 1;
+        int days = Math.toIntExact(((afterTime - startTime) / (1000 * 3600 * 24))) + 1;
         return days;
     }
